+++ README.md
@@ -1,10 +1,2 @@
-## タスク1で編集したファイル
-```
--src
-    -app
-        -page.tsx
-    -components
-        -Header.tsx
-        -Section.tsx
-```
-Header.tsxでヘッダーを作成し、Section.tsxで主題のを作成し、それをpage.tsxで読み込むというふうになっています。
+## タスク3
+コンポーネント化の見直しをしたため追加は少ないですが変更点が多くなってます
