+++ src/components/ApplicantListHeader/ApplicantCountDisplay.tsx
@@ -1,0 +1,7 @@
+const ApplicantCountDisplay: React.FC = () => (
+  <div className="mt-2 text-[#666] text-[14px]">
+    全XXX件中1から50件を表示
+  </div>
+);
+
+export default ApplicantCountDisplay;
