+++ src/components/ApplicantListHeader/ApplicantTitle.tsx
@@ -1,0 +1,7 @@
+const ApplicantTitle: React.FC = () => (
+  <h1 className="text-[2.2rem] font-bold m-0 mb-5">
+    求職者管理
+  </h1>
+);
+
+export default ApplicantTitle;
